refact: modalDelete component
